+++ examples/noise.rs
@@ -1,11 +1,14 @@
 use {
     bevy::{
-        prelude::{App, Assets, Commands, Deref, Handle, Res, ResMut, Resource, StartupStage},
+        prelude::{
+            App, Assets, Commands, Deref, Handle, IntoSystemConfig, Res, ResMut, Resource,
+            StartupSet,
+        },
         reflect::TypeUuid,
         DefaultPlugins,
     },
-    bevy_oddio::{frames::Stereo, output::AudioSink, Audio, AudioApp, AudioPlugin, ToSignal},
-    oddio::Signal,
+    bevy_oddio::{output::AudioSink, Audio, AudioApp, AudioPlugin, ToSignal},
+    oddio::{Sample, Signal},
 };
 
 #[derive(TypeUuid)]
@@ -15,7 +18,7 @@ struct Noise;
 struct NoiseSignal;
 
 impl Signal for NoiseSignal {
-    type Frame = Stereo;
+    type Frame = [Sample; 2];
 
     fn sample(&self, _interval: f32, out: &mut [Self::Frame]) {
         for out_frame in out {
@@ -39,9 +42,9 @@ fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
         .add_plugin(AudioPlugin::new())
-        .add_audio_source::<2, _, Noise>()
+        .add_audio_source::<_, Noise>()
         .add_startup_system(init_assets)
-        .add_startup_system_to_stage(StartupStage::PostStartup, play_noise)
+        .add_startup_system(play_noise.in_base_set(StartupSet::PostStartup))
         .run();
 }
 
@@ -57,7 +60,7 @@ fn init_assets(mut commands: Commands, mut assets: ResMut<Assets<Noise>>) {
 
 fn play_noise(
     mut commands: Commands,
-    mut audio: ResMut<Audio<Stereo, Noise>>,
+    mut audio: ResMut<Audio<[Sample; 2], Noise>>,
     noise: Res<NoiseHandle>,
 ) {
     let handle = audio.play(noise.clone(), ());