+++ examples/sine.rs
@@ -1,6 +1,9 @@
 use {
     bevy::{
-        prelude::{App, Assets, Commands, Deref, Handle, Res, ResMut, Resource, StartupStage},
+        prelude::{
+            App, Assets, Commands, Deref, Handle, IntoSystemConfig, Res, ResMut, Resource,
+            StartupSet,
+        },
         DefaultPlugins,
     },
     bevy_oddio::{
@@ -16,7 +19,7 @@ fn main() {
         .add_plugins(DefaultPlugins)
         .add_plugin(AudioPlugin::new())
         .add_startup_system(init_assets)
-        .add_startup_system_to_stage(StartupStage::PostStartup, play_sine)
+        .add_startup_system(play_sine.in_base_set(StartupSet::PostStartup))
         .run();
 }