+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-13-07
+
+### Added
+
+- `ArrayLength` and `AsArray` traits. You should really only use `AsArray`.
+
+### Changed
+
+- `AudioOutput<N, F>` to `AudioOutput<F>` 
+- `add_audio_source` and `play_queued_audio` removed its `const N: usize` generic.
+- `frame_n` requires `AsArray` as a bound instead of `FromFrame` and `const N: usize`.
+
+### Removed
+
+- `Mono` and `Stereo`. Use `[f32; 1]` and `[f32; 2]`, respectively.
+- `FromFrame`. Use the `AsArray` trait for custom frames.
+
 ## [0.3.0] - 2022-11-13
 
 ### Added
@@ -64,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Released `bevy_oddio` 0.1 🎉
 
-[Unreleased]: https://github.com/harudagondi/bevy_oddio/compare/v0.2.0..HEAD
+[Unreleased]: https://github.com/harudagondi/bevy_oddio/compare/v0.4.0..HEAD
+[0.4.0]: https://github.com/harudagondi/bevy_oddio/compare/v0.3.0..v0.4.0
 [0.3.0]: https://github.com/harudagondi/bevy_oddio/compare/v0.2.0..v0.3.0
 [0.2.0]: https://github.com/harudagondi/bevy_oddio/compare/v0.1.0..v0.2.0
 [0.1.0]: https://github.com/harudagondi/bevy_oddio/releases/tag/v0.1.0